requirements.txt
telebot
pyTelegramBotAPI
pandas
python-binance==0.7.9
twisted 